Created new Util class
